+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -93,5 +93,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-19T07:49:13.924Z")
+    created: Optional[datetime] = parse_datetime("2025-05-19T10:31:02.146Z")
     r"""The date and time the resource was created"""