+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-19T07:49:11.066Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-19T10:30:59.370Z")
     r"""The date and time the resource was last updated"""