+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVKR9F5HGV4XC632YGHPDN4E"
+        "01JVM1HQVMEY30KG3GECJ1JB48"
     )
     r"""The id of the resource"""