+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -86,7 +86,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVKR9F5J5GP4SJY8PE0TCV3G"
+        "01JVM1HQVRY39CXSHBY4Y5FBMK"
     )
     r"""The id of the resource"""