+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVKR9F5MFC5GEX9AC2J26GV0"
+        "01JVM1HQVTXE6YDR2D6NSD1WVR"
     )
     r"""The id of the resource"""