+++ packages/orq-rc/RELEASES.md
@@ -1259,3 +1259,13 @@ Based on:
 - [python v3.6.0-rc.11] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.11] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.11 - packages/orq-rc
+
+## 2025-05-19 10:32:39
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.546.2 (2.604.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.12] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.12] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.12 - packages/orq-rc